+++ include/hybrid_BC/main.cuh
@@ -7,4 +7,4 @@
 #include "util.cuh"
 #include "kernels.cuh"
 
-std::map<int, float> compute_bc(std::vector<std::pair<int, int> > edges, int k, bool approx,	int device );
+std::map<int, double> compute_bc(std::vector<std::pair<int, int> > edges, int k, bool approx,	int device );
